requirements.txt
anyio==3.6.2
appdirs==1.4.4
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
cssselect==1.2.0
Django==4.1.7
dnspython==2.3.0
exceptiongroup==1.1.1
fake-useragent==1.1.3
fastapi==0.92.0
h11==0.14.0
hiredis==2.2.2
httptools==0.5.0
idna==3.4
importlib-metadata==6.1.0
lxml==4.9.2
numpy==1.24.2
outcome==1.2.0
pandas==1.5.3
parse==1.19.0
pycparser==2.21
pydantic==1.10.5
pyee==8.2.2
PyJWT==2.6.0
pymongo==4.3.3
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-html==0.10.0
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
starlette==0.25.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
w3lib==2.1.1
watchfiles==0.18.1
websockets==10.4
wsproto==1.2.0
zipp==3.15.0
